+++ server/lara-course/resources/views/users/table.blade.php
@@ -20,6 +20,7 @@
                 <td><a class="text-secondary font-weight-bold text-capitalize" href="mailto:{{$user->email}}"> {{ $user->email }}</a></td>
                 <td>{{ $user->view_count }}</td>
                 <td>{{ $user->role['name'] }}</td>
+                @if(Auth::user()->id == $user->id || Auth::user()->role_id < 3)
                     <td>
                         {!! Form::open(['route' => ['users.destroy', $user->id], 'method' => 'delete']) !!}
                         <div class='btn-group'>
@@ -29,6 +30,7 @@
                         </div>
                         {!! Form::close() !!}
                     </td>
+                @endif
             </tr>
         @endforeach
         </tbody>
